Project_Starter_Files-Building_a_Metrics_Dashboard/app/backend/requirements.txt
Flask==2.1.0
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo
prometheus-client==0.12.0
Flask_OpenTracing==1.1.0
jaeger-client==4.3.0
opentelemetry-api
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation
opentelemetry-sdk
prometheus-flask-exporter==0.18.1
